requirements.txt
# web
requests

# infra
python-dotenv

# db
clickhouse_driver